Remove the AOC Service
