+++ Guppi.Application/Configurations/AdventOfCodeConfiguration.cs
@@ -1,7 +1,7 @@
 using Guppi.Application;
 using Guppi.Application.Attributes;
 
-namespace Guppi.Infrastructure.Services.AdventOfCode
+namespace Guppi.Application.Configurations
 {
     class AdventOfCodeConfiguration : Configuration
     {
